requirements.txt
asgiref==3.8.1
django-cors-headers==4.3.1
dj-database-url==2.1.0
Django==5.0.6
djangorestframework==3.15.1
Pillow==10.4.0
pytz==2024.1
sqlparse==0.5.0
whitenoise==6.7.0